@@ -1,9 +1,9 @@
 package kubernetes
 
 import (
+	"errors"
 	"fmt"
 	"github.com/imdario/mergo"
-	"github.com/pkg/errors"
 	restclient "k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
