@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"encoding/xml"
 	log "github.com/sirupsen/logrus"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 	"net/http"
 	"time"
 )
